A hacky fix for specializing methods from extensions
If we don't find the generic we expect in the first pass during IR specialization, then we check for the special case where we are trying to specialize something from a generic extension, using the type being extended.
We assume that the generic parameter lists match (that part is the huge hack), and collect the arguments as if they were for the extension instead of the type.

This will break when/if we ever have generic extensions with parameter lists that don't match the type being extended.