Fix a bug in IR use-def information
The basic problem here is that when unlinking an `IRUse` from the linked list of uses, there were several cases where I was failing to set the `prevLink` field of the next node to match the `prevLink` field of the node being removed. That doesn't show up when walking the linked list of uses forward, but it breaks it whenever you have subsequent unlinking operations.

This change fixes the bugs of that kind I could find, and also adds a debug validation method to try to avoid breaking it again. I also made more access to `IRUse` go through accessor methods rather than using fields directly, to try to avoid this kind of error. I stopped short of making anything `private`, because I tend to find that it creates more hassles than it avoids.

A few other fixes along the way:

- Made the `List<T>` type default-initialize elements when you resize it. I hadn't realized we weren't doing that.

- Add a standalone `dumpIR(IRGlobalValue*)` so help when debugging issues.