Fix infinite loop in semantic parsing
The code for parsing semantics was looking for a fixed set of tokens to terminate a semantic list, rather than assuming that whenever you don't see a `:` ahead, you probably are done with semantics. This meant that you could get into an infinite loop just with simple mistakes like leaving out a `;`.

This change fixes the parser to note infinite loop in this case, and adds a test case to verify the fix.